XW-30 code unification after XW-474 merge
