+++ front/scripts/main/routes/MainPageCategoryRoute.ts
@@ -22,7 +22,7 @@ App.MainPageCategoryRoute = Em.Route.extend(App.MainPageRouteMixin, App.MetaTags
 		error(error: any): boolean {
 			// Status comes from ArticlesApiController::getList in MediaWiki
 			// and code comes from MercuryApiController in MW and server side code in Mercury app
-			if (error && error.status === 404 || error.code === 404) {
+			if (error && (error.status === 404 || error.code === 404)) {
 				this.controllerFor('application').addAlert({
 					message: i18n.t('app.curated-content-error-category-not-found'),
 					type: 'warning',
