+++ typings/mercury/mercury-server.d.ts
@@ -87,4 +87,3 @@ interface ServerData {
 	apiBase: string;
 	environment: string;
 }
-
